Run swig from the main setup.py when building the renamer's XML instead of from build_renamers.py.  This makes it easy to ensure that the same flags are used.  Also, use the new swig flag to generate the XML after the python language module has run, rather than instead of. This gives us more info in the xml output.
Updated build_renamers.py to expect the xml file to already be
generated, and also allow it to run with a xml doc tree from libxml2
rather than the xml.sax module.  This makes the extraction of info
much cleaner and logical since we usually need info from child or
grandchild nodes.