Append default extension before showing file save dialog, not after.
Appending the extension after the dialog was hidden is a bad idea as it
misleads the user by using a different file name from the one shown in the
dialog. It is also dangerous as it bypassed wxFD_OVERWRITE_PROMPT check.

So just append the default extension to the initial file name if it doesn't
have any but don't modify the file name once it was accepted by user.

Closes #11256.