Fix selection event sending for Ctrl-click in generic wxDataViewCtrl.
Ctrl-clicking on the only currently selected item in wxDataViewCtrl crashed
because we dereferenced first element of the empty m_selection array when
generating the selection change event.

Instead, send the event for the current item, this should be more consistent
with the native versions behaviour and at the very least doesn't crash.

Closes #11684.