Delete pending objects in wxApp::ProcessPendingEvents() and not ProcessIdle().
Move DeletePendingObjects() call from ProcessPendingEvents() to ProcessIdle()
to ensure that we delete the objects marked for destruction even if the
application is sitting in a tight OnIdle() loop, i.e. if the idle event
handler keeps requesting more events.

Also make sure that the event loop terminates if its OnExit() was called even
if the idle event handler continues to request more events.

Closes #12424.