Replace a name with Latin-1 characters with ASCII equivalent.
The file wxchar.h currently doesn't have UTF-8 encoding but contained
non-ASCII 'å' character which prevented any changes to it from being
committed. Instead of converting the file to UTF-8 (which would be better but
just might create some unforeseen problems) simply use the ASCII spelling of
the name, as it's already done in several other files.