Work around an intermittent crash in GSocket under Mac.
The socket descriptor may somehow be invalid when an input timeout
notification is received.

This is almost certainly not the right fix but the code has completely changed
in 2.9 so it doesn't make much sense to spend time on fixing it properly here.
But adding this check seems to be enough to fix the crash, see #12264.

Closes #8904.