Add USE_ANIMATION build option to config.bkl file.
This option was used by multilib.bkl but never defined, so running bakefile
failed.

Do add it and define its default value as 1 for the formats not supporting
conditions, such as MSVC IDE.

Also mention the need to do this in the tech note explaining how to add new
libraries.